Run tests parallel
Speeds up test runner from ~10.5 seconds to ~2.8 seconds, which is not a
bad win.
