+++ auth/kerberos/krb_unix.go
@@ -1,5 +1,7 @@
 //go:build !windows
 
+// UNIX Kerberos support, using jcmturner's pure-go implementation
+
 package kerberos
 
 import (
@@ -13,11 +15,6 @@ import (
 	"github.com/jcmturner/gokrb5/v8/credentials"
 	"github.com/jcmturner/gokrb5/v8/spnego"
 )
-
-/*
- * UNIX Kerberos support, using jcmturner's pure-go
- * implementation
- */
 
 // GSS implements the pq.GSS interface.
 type GSS struct {
